ENH: Meal composing templates
Added a template for new meal view for composing and
viewing details of a meal.

Add ingredient card contains a search + result field
with a preview for the selected ingredient and add
ingredient form. Ingredient preview contains hardcoded
sample values for now.

Ingredient preview contains a macronutrient calorie pie
chart, the description of the ingredient and space for
badges.

+ gitignore ignores data dir only in the root dir