ENH: Profile energy calculation
Added a method for calculating the estimated energy
requirement (EER). The EER is calculated and saved when
profile model save method is called.

Altered profile model:
- changed all fields to non-nullable
- added weight and energy_requirement fields

The weight field was also added to the profile form.

A data migration was added to calculate and save the
energy requirement value for existing profiles.

+ removed squashed migrations.