ENH: MealView recommendations and recommendation serializer
MealView was added to display micronutrient
recommendations.
Meal View is tasked with displaying recommendation and
intake data for a user profile organized by nutrient types.

A new serializer was added for serializing
IntakeRecommendations taking into account profile weight
and energy requirement.
The serializer shows, in addition to IntakeRecommendation
fields not used for selection, the related nutrient's
id, name, types, unit symbol, energy per unit.
It also selects the upper limit of a recommendation
(the UL value or upper limit of AMDR), the
recommendation's displayed amount (rounded to the first
decimal place) and profile_amount_min for the current
profile.

The micronutrients.html template was edited to incorporate
context data.
A template was added for micronutrient table rows.

+ Styling changes: conditional progressbar colors.
+ logged_in_client fixture: client with the user from the
`user` fixture logged in.