FEAT: Meal Tracking
Implemented the meal tracking feature that allows users to register the
foods eaten by them in a particular day and view the resulting nutrient
intake. Users can select the meal by date and add, remove, and change
the amount of ingredients in a meal.

Models
------
Repurposed the `Meal` model to represent the foods eaten by the user in
a given day and added the `MealIngredient` model to represent the amount
of an ingredient in a meal.
The `Meal` model no longer has a `name` field, and the `user` field was
renamed to `owner` and now references the `Profile` model instead of the
`User` model.
Added the `get_intakes()` method for calculating the amount of each
nutrient in the meal.

Added the new `parent_nutrient` field to the `NutrientType` model and
updated the `populatenutrientdata` command accordingly. Modified the
models' save method to prevent hierarchies: A NutrientType can't have a
`parent_nutrient` with a type that also has a non-null
`parent_nutrient`.

Views
-----
Modified the MealView and the IngredientPreviewView, and implemented new
API views: MealIngredientView, MealIngredientDetailView,
CurrentMealView, CurrentMealRedirectView and MealNutrientIntakeView.

The `MealView` is responsible for managing the user's current meal, in
particular creating new meals if necessary and checking for expiration
of the current meal. Besides that, the `MealView` is a TemplateView that
handles the layout of the page where the user views and modifies their
meal. The data is lazy loaded through requests to API views.
Renamed the associated url from `current-meal` to `meal`.

The `MealIngredientView` allows the user that owns the meal to list
a meal's ingredients and add ingredients to that meal. The meal
ingredients are listed in reverse order to simultaneously allow lazy
loading of paginated data, when scrolling down, and appending a newly
added ingredient to the top of the list.
The view uses a template for the table displaying the meal ingredients
and a template for a single row of that table for POST requests.

The `MealIngredientDetailView` is for retrieve, update and destroy
operations on `MealIngredient` entries. The view uses two templates: the
default table row template and the table row form template for GET
requests that included `template=form` in their query params.

If the request was successful, the `MealIngredientView` and
`MealIngredientDetailView` response includes the `mealComponentsChanged`
HTMX event header to notify the page that the displayed nutrient
intakes need to be updated.

The CurrentMealView is for setting by date and retrieving the current
meal. The response for a successful POST request to the view includes
the `currentMealChanged` HTMX event header to notify the page that the
displayed nutrient intakes and meal ingredients need to be updated.
If the `TemplateHTMLRenderer` is used, the view adds a datetime object
of the date to the response data to allow formatting in the template.

The CurrentMealRedirectView redirects requests to views by selecting the
current meal's id. Currently, this is used to redirect to the
MealIngredientView and MealNutrientIntakeView.

The MealNutrientIntakeView lists the nutrient intakes of the meal.
The view uses a complex queryset that only includes the nutrients with
an associated `IntakeRecommendation` except nutrients specified in the
view's `no_recommendations` attribute. The view allows specifying
nutrients for which to skip recommendation entries with the `dri_type`
set to 'AMDR' to prevent duplications. The returned nutrients are
grouped and ordered according to the `display_order` attribute. The
recommendations are filtered to match the current user's profile.
The returned data, in addition to nutrient information, includes
associated intake recommendations, calculated intakes from the meal and
child nutrients through the `NutrientType.parent_nutrient` relation.
The view uses the `IsMealOwnerPermission` to prevent users that don't
own the meal from viewing the intakes.

The IngredientPreviewView (previously IngredientPreview) was moved to
htmx.py.
The view appends the current meal's id to the response data if the
renderer used is a `TemplateHTMLRenderer`. This is used to point the
'add ingredient' form to the correct meal.

Current Meal System
-------------------
Implemented a system holding the id of the meal currently being
interacted with by the user in the user's session. This helps
synchronize the meal edited by API views and prevents the need for the
user to select the desired meal every time the user visits the meal
page.
It is possible to set the time from the last interaction with the meal
after which the meal expires using the MEAL_EXPIRY_TIME setting. If the
current meal is expired, a request to the MealView will set the current
meal to one with the date of the request.

The `clearemptymeals` Command
-----------------------------
Implemented a command that removes Meal entries without any related
MealIngredients and the meal isn't any session's current meal. This is
needed because meals are created when they are selected by date in the
CurrentMealView or MealView without considering if the user made any
changes to the meal.

Permissions
-----------
Implemented the `IsMealOwnerPermission` allowing only the owner of a
meal to view or modify it, and the `IsMealOwnerComponentPermission`
object permission allowing to view or modify the object if the user owns
its related meal.

Serializers
-----------
Replaced the `ProfileIntakeRecommendationSerializer` with the newly
implemented `NutrientIntakeSerializer` and `RecommendationSerializer`.
The `RecommendationSerializer` class handles representation of
recommendations and intakes associated with them. The intake information
is provided through the `meal` context var.
The `NutrientIntakeSerializer`, in addition to the nutrient fields, has
a nested RecommendationSerializer. The list serializer groups the by
types and can be filtered and ordered. It also displays nested
NutrientIntake representation for parent_nutrient relations of the
types. The serializer data includes the intake info in the same manner
as the `RecommendationSerializer` to allow displaying the intakes even
if the nutrient has no related recommendations.

Implemented the `CurrentMealSerializer` for displaying and creating
meals. It uses get_or_create() in its create method.

Added the `MealIngredientSerializer`.