FEAT: Recipe system
** MealComponent repurpose **

Repurposed the MealComponent, MealComponentIngredient and
MealComponentAmount models for the recipe system.

Recipe:
- Renamed the MealComponent model to 'Recipe.'
- Changed the `user` field into the `owner` field and changed the
reference to the `Profile` model.
- Added a many-to-many relationship with ingredients through the
`RecipeIngredient` (formerly MealComponentIngredient) model.
- Added the `slug` field and implemented the `Recipe.get_slug()` method
for generating the recipe's slug from its name. The method is always
called when saving an instance.
- Added to the model owner-name and owner-slug unique together
constraints.
- Made the `final_weight` field optional and added a MinValueValidator
to the field

RecipeIngredient:
- Renamed the MealComponentIngredient model to 'RecipeIngredient.'
- Added a MinValueValidator to the `amount` field.
- Removed the unique_recipe_ingredient constraint.

MealRecipe:
- Renamed the MealComponentAmount model to 'MealRecipe.'
- Added MinValueValidator to he `amount` field.
- Removed the unique together constraint for the `meal` and `recipe`
fields.

Reimplemented the Meal model's `get_intakes()` method to include intakes
from recipes and the Recipe model's nutritional_value() to use a single
query and return the values per gram.
Added a new Recipe.get_intakes() method to return the recipe's
nutritional value per 100g.
Implemented the `Recipe` model's `calories` property mirroring the
`Ingredient` model's `calories` property.

** Views **

Implemented the RecipeEditView for editing recipes in a similar manner
to the MealView.

Implemented a view set base class `ComponentCollectionViewset` for
component collection models (MealIngredient, MealRecipe and
RecipeIngredient) to reuse code from the MealIngredientView and
the MealIngredientDetailView. Generalized the related templates.
Generalized the relevant permissions and created a new permission
`IsOwnerPermission` for views where the collection permissions don't
apply.
Implemented the `ModelCollectionRouter` for the view sets to match the
old url patterns.
Implemented and unified the structure of component collection preview
views and serializers.
Added the preview_url field to the IngredientSerializer and
the RecipeSerializer. The url in the case of the IngredientSerializer
is dynamically selected based on the `target` query param.
Added the `url` field to the component collection serializers
pointing to their respective detail views.

Generalized the NutrientIntakeView for use with both recipes and meals.

Moved the mixins to a separate module and generalized the
`MealEditMixin` and `RecipeEditMixin` into the `HTMXEventMixin` that
uses the events in `htmx_events` view attribute in the HX-Trigger
header.

Implemented the `RecipeViewSet` to add list, create, retrieve, update
and destroy endpoints for the recipe model.
The viewset functionality includes redirects on create and destroy.
Added and updated related templates.

Added a template view to the url conf allowing the user to select a
recipe to modify (select recipe page). The page lazy-loads the recipes
via a request to the RecipeView.
The page contains a form for creating a new recipe.

Added HTMX-driven tabs in the search component of the meal page to allow
searching for both ingredients and recipes. Implemented the
`MealComponentTabView` to manage the active tabs.
The meal's recipes are also lazy loaded to the meal page's component
list using a new endpoint to the `CurrentMealRedirectView` for recipes.

Separated the views from the htmx.py module to multiple smaller modules
based on functionality:
- base_views.py - View base classes intended for use only through
inheritance.
- api_views.py - Core views that didn't need a new module.
- current_meal_views.py - Views related to the 'current-meal'
functionality.
- hypermedia_views.py - Views that only render hypermedia based on the
state of the application.
- meal_views.py - Views related to the `Meal` model.
- recipe_views.py - Views related to the `Recipe` model.

Separated the related tests accordingly.

Updated the sidebar to include separate Meal and Recipe links.

+ TST: Updated the `clearemptymeal` command's 'number of queries' test
to include the deletion of related MealRecipe entries.