FEAT: Weight measurement system
Change the way the profile's weight is handled by the app. Now the users
input weight measurements. The profile's weight value is the average
value of all measurements within a week from the last one.

Add the `WeightMeasurement` model to represent a measurement of the
person's, represented by the `profile`, weight.

Add the `current_weight` property to the `Profile` model representing
the weight based on the average of the profile's weight measurements.

Modify the `Profile` model's save method to create a `WeightMeasurement`
entry if the profile is being created or the `add_measurement` argument
is True, and to set the profile's weight to the value of the
`current_weight` property if the `recalculate_weight` argument is True.
Add the `update_weight()` method to the Profile model that sets the
profile's weight to the value of the `current_weight` property.

Add the `WeightMeasurementViewSet` with the related serializer and the
'weight measurements' page to allow CRUD operations on
WeightMeasurements.

Add the new `unit` field to the `WeightMeasurementSerializer` to allow
specifying if the weight value is in kilograms or pounds. Modify the
serializer to convert the value to kilograms before saving.
Add a select field for units to weight measurement forms.

+ Enable the response-targets HTMX extension
+ Rename the test_profile.py module to 'test_profile_models.py'
+ Turn off timezone support. Timezone support is not needed for the
functioning of the application. Instead of increasing complexity to
accommodate user date inputs, timezone support is turned off.
+ Move the pounds_to_kilograms() function from main/forms.py to util.py.
+ Modify the settings pages' templates to match the navigation
scheme of:

settings {
  account settings {
    change password,
    reset password
  },
  profile settings {
    profile information,
    weight measurements
  }
}

Adjust the templates to unify appearance.