ENH: Async profile information form
Add the ProfileApiView to allow users to perform create, update and
retrieve operations on their profile.

Change the main ProfileView to a class-based view.
Remove the use of the ProfileForm from the view. The API view
will handle the form functionality.
Change the profile page's url name to 'profile-information' and path to
'profile_information.'
Add support for the redirect url field `next` to the view.

Move the automatic unit conversion script from profile.html to index.js.

Remove unused ProfileForm

Move centimeters_to_feet_and_inches() to util.py

Replace the placeholder personal information modal with a functional one
that uses the ProfileApiView.
Change the modal's label to 'Profile Information.'

Introduce the `profile_required` decorator that checks if the user has
completed profile creation, and if not, redirects to the profile
creation page in a similar manner to the `login_required` decorator.
Apply the decorator to pages that require a profile to function
properly.