FEAT: Implement the Nutrient tracking system
The new feature allows users to select nutrients to pay special
attention to. Intake charts for the selected nutrients will be displayed
in the dashboard.

Profile model
-------------
- Add intakes_by_date method that calculates the intake of a nutrient by
date for the profile.
- Add the tracked_nutrients m2m field.

Views
-----
- Add the LastMonthIntakeView which provides data about the profile's
intakes of a nutrient, within 30 days before the request, by date.
The data is formatted by the ByDateIntakeSerializer to be compatible
with frontend chart functions.
- Add the TrackedNutrientViewSet which allows users to perform list,
create and destroy operations on tracked nutrients. Update the dashboard
page template to lazy-load and auto-refresh the tracked nutrient charts
using the view set. Replace the placeholder table in dashboard 'tracked
nutrient' modal with a functional one.

Frontend
--------
- Add the monthIntakeChart which draws intake charts based on data from
the LastMonthIntakeView.
- Add the fetchLastMonthIntake function that makes a request to the
LastMonthIntakeView and calls monthIntakeChart with the response data.
- Replace the static placeholder charts with ones utilizing the new
functionality.