FIX: Fix repeating rows in meal's component list
On the meal page, after changing the date, the components of the meal
in the selected date were appended to the bottom of the list instead
of replacing the old list.

To fix this, the htmx related attributes were configured so that the
response from the MealIngredientView would replace the old list and the
response from the MealRecipeView would be appended after that.