ENH: Create custom generics and model mixins
Create custom versions of the GenericView, GenericViewSet, ModelViewSet
and model mixins utilizing the new functionality to better match the
project's use case and reduce repetition of common patterns in view
implementations.
The changes effect behavior only if the view selected the template
renderer.

Using the new generics and mixins allows to:
- Define a template map instead of overwriting the
`get_template_names()` method.
- Add additional context through the `get_template_context()` method.
- Define response headers for unsafe HTTP methods / actions depending
on whether the operation was successful or not using the
`get_success_headers()` and 'get_fail_headers()' methods.
- Access the value of the request's 'template' query param, if one was
included in the url, through the `template` property.

+ Move mixin module to the views package.