ENH: Add recommendation intake properties
Add the set_up() method properties related to nutrient intake to the
IntakeRecommendation model.
The set_up() method sets the recommendations profile and
intake. This allows to access some recommendation values as properties

displayed_amount: The displayed amount for a recommendation.
over_limit: Whether the nutrient's intake is higher than or equal to the
  recommended upper limit.
progress: The ratio of the nutrient intake to the recommended nutrient
  intake in percent.