ENH: Create first_instance_of template tag
The tag behaves the same way as the built-in `firstof` tag, except when
assigning to a context variable using the 'as' keyword, the selected
value is not converted to a string.

This allows priority selection from a iterable in a template.
For example, selecting from an unknown number of recommendations for a
nutrient.