ENH: Remove intake serializers template behavior
Modify the NutrientIntakeSerializer and the Recommendation serializer
to remove features that were implemented for use with a template.
The new implementation of the NutrientIntakeView doesn't use the
serializers to render template responses.

Remove the custom list serializers.

Change the RecommendationSerializer method fields to default fields
utilizing the new intake related properties of the IntakeRecommendation
model.

Make the context requirements of the serializers more strict: Require
`request` and optionally `intakes`.

+ Add id field to NutrientTypeSerializer
+ NutrientIntakeSerializer property fields
+ Change NutrientIntakeSerializer child_type field to id instead of
name