MAINT: Refactor serializers
Remove serializers and fields that were needed only for rendering
templates. The views inheriting from new generics no longer use
serializers for rendering templates.

Modify the create methods to either require providing missing fields
in the save method args or providing the request in the context.
The request in context is usually required if it is used elsewhere in
the serializer.

Update ByDateIntakeSerializer docstring to describe the context vars
used by the serializer.

Remove MissingContextError class (the class is no longer used).

Rename TrackedNutrientSerializer `name` field to `nutrient_name`.

Change to IngredientSerializer's superclass to `ModelSerializer`,
because the url field was removed.

Remove date formatting statements in the WeightMeasurementSerializer,
as they were only needed for compatibility with form fields.