FIX: Deleted meal in session breaks meal page
If a meal, that was a sessions current meal, was deleted the current
meal views would stop working. Users don't have the ability to
remove meals, but a meal might be deleted through admin action or
cascade delete.

Now the main `MealView` will create a new meal if the meal in the
session does not exist.