ENH: Change weight time to date & value to float.
Change the WeightMeasurement time field to a `DateField` (renamed to
date) and the value field to a `FloatField`.
Remove time & profile unique together constraint.

The exact time of the weight measurement is not really needed and using
just the dates avoids messing with timezones.
Using float values allows to more easily show trends.

+ Remove unused imports.
+ Change step size to 0.1 for value fields in forms.