ENH: Adjust intake view api fields and add renderer
Add BrowsableAPIRenderer to the NutrientIntakeView.

Remove detail nutrient data fields from the NutrientIntakeSerializer.
Add a url to the nutrient detail location instead.

Replace the displayed_amount field in the RecommendationIntakeSerializer
with the amount_min and amount_max fields inherited from the
RecommendationSerializer.

Add demographic data fields to the RecommendationSerializer. The data is
useful in cases where the api is accessed by a user without a profile.

Make the NutrientIntakeView queryset prefetch change based on whether
the template renderer was used to match the serializer fields.