FIX: Incorrect intakes for multiple occurrences of recipes
Meal calories and intakes properties returned incorrect value in cases
where the meal had multiple occurrences of the same recipe and the
recipe had a null final weight.

The queryset calculating the recipe weights summed the weights for each
occurrence of the recipe resulting in the correct weight being
multiplied by the number of occurrences.

The queryset was modified to calculate the values using a filtered
Recipe queryset instead of the Meal.recipes related manager.