FIX: Double scroll on content
The content on each page after scrolling to the bottom could be scrolled
down again shifting the page down. If the page was scrolled to the top,
some of the content is cut off until the page is scrolled up again.

This was caused by the top margin on the content wrapper. The margin was
replaced with padding.