FIX: create_compound_nutrient_amounts PostgreSQL error
When testing on a PostgreSQL database, the function caused an error,
because the query retrieving compound nutrients returned duplicated
instances for each component of the nutrient.
This later caused issues because there were duplicated
IngredientNutrient instances in the 'bulk_create' statement with
update_conflicts set to True, which is not allowed in PostgreSQL.

The query was changed to one that doesn't return duplicates.