FIX: Component collection view # of queries test
The `BaseModelCollectionViewsetTests.test_list_num_queries` test used
an ingredient fixture in the data instead of a component fixture defined
by a concrete subclass. This broke the test for the MealRecipeViewSet.

An additional fixture `extra_component` was added to the concrete
classes, which return another instance of the component model that does
not have a relation with the collection instance.
The fixture replaced the use of the ingredient fixture.