FIX: Debug toolbar error using non dev settings
The urls conf for the project caused an error when adding debug toolbar
locations. The urls were added only if DEBUG was set to True but did not
check if the debug_toolbar was included in INSTALLED_APPS (which was the
case only in the dev settings).

Added a check ensuring that debug_toolbar is installed
before adding the locations.